- moved mixins to @/mixins - added loading state to Download Receipt button
