+++ src/components/common/applicant-info-2.vue
@@ -206,7 +206,7 @@ import { Component, Vue, Watch } from 'vue-property-decorator'
 import newReqModule, { NewRequestModule } from '@/store/new-request-module'
 import paymentModule from '@/modules/payment'
 import ApplicantInfoNav from '@/components/common/applicant-info-nav.vue'
-import NameRequestMixin from '@/components/mixins/name-request-mixin'
+import { NameRequestMixin } from '@/mixins'
 
 @Component({
   components: {
