+++ src/components/new-request/main-container.vue
@@ -24,7 +24,7 @@
 <script lang="ts">
 import { Component, Mixins } from 'vue-property-decorator'
 
-import DisplayedComponentMixin from '@/components/mixins/displayed-component-mixin'
+import { DisplayedComponentMixin } from '@/mixins'
 import newReqModule from '@/store/new-request-module'
 
 @Component({})
