+++ src/components/payment/reapply-modal.vue
@@ -35,12 +35,8 @@ import * as PaymentTypes from '@/modules/payment/store/types'
 import * as FilingTypes from '@/modules/payment/filing-types'
 import * as Jurisdictions from '@/modules/payment/jurisdictions'
 import { PaymentAction } from '@/enums'
-import PaymentMixin from '@/components/payment/payment-mixin'
-import PaymentSessionMixin from '@/components/payment/payment-session-mixin'
-import NameRequestMixin from '@/components/mixins/name-request-mixin'
-import DisplayedComponentMixin from '@/components/mixins/displayed-component-mixin'
+import { NameRequestMixin, PaymentMixin, PaymentSessionMixin, DisplayedComponentMixin } from '@/mixins'
 import { getBaseUrl } from './payment-utils'
-import newReqModule from '@/store/new-request-module'
 
 @Component({
   components: {
