+++ src/components/payment/refund-modal.vue
@@ -44,12 +44,10 @@
 <script lang="ts">
 import { Component, Mixins, Watch } from 'vue-property-decorator'
 import PaymentModule from '@/modules/payment'
-import PaymentMixin from '@/components/payment/payment-mixin'
-import PaymentSessionMixin from '@/components/payment/payment-session-mixin'
 import RefundSummary from '@/components/payment/refund-summary.vue'
 import * as PaymentTypes from '@/modules/payment/store/types'
 import { NrAction } from '@/enums'
-import NameRequestMixin from '@/components/mixins/name-request-mixin'
+import { NameRequestMixin, PaymentMixin, PaymentSessionMixin } from '@/mixins'
 import NewReqModule from '@/store/new-request-module'
 import { sleep } from '@/plugins'
