+++ src/mixins/index.ts
@@ -1,0 +1,7 @@
+export { default as PaymentMixin } from './payment/payment-mixin'
+export { default as PaymentSessionMixin } from './payment/payment-session-mixin'
+export { default as CommonMixin } from './common-mixin'
+export { default as DateMixin } from './date-mixin'
+export { default as DisplayedComponentMixin } from './displayed-component-mixin'
+export { default as NameRequestMixin } from './name-request-mixin'
+export { default as NrAffiliationMixin } from './nr-affiliation-mixin'
