+++ tests/unit/applicant-info-1.spec.ts
@@ -1,6 +1,5 @@
 import ApplicantInfo1 from '@/components/common/applicant-info-1.vue'
 import { createLocalVue, shallowMount, mount } from '@vue/test-utils'
-import newReqModule from '@/store/new-request-module'
 import Vuetify from 'vuetify'
 import $intJurisdictions from '@/store/list-data/intl-jurisdictions'
 import $canJurisdictions from '@/store/list-data/canada-jurisdictions'
